+++ core/core.pl
@@ -3402,7 +3402,7 @@ create_logtalk_flag(Flag, Value, Options) :-
 % versions, 'rcN' for release candidates (with N being a natural number),
 % and 'stable' for stable versions
 
-'$lgt_version_data'(logtalk(3, 29, 0, b03)).
+'$lgt_version_data'(logtalk(3, 29, 0, b04)).
 
 
 
@@ -11559,6 +11559,8 @@ create_logtalk_flag(Flag, Value, Options) :-
 	'$lgt_comp_ctx_exec_ctx'(Ctx, ExCtx),
 	'$lgt_compile_body'(Pred, TPred, DPred, Ctx).
 
+% deprecated Prolog built-in predicates
+
 '$lgt_compile_body'(not(Pred), TPred, DPred, Ctx) :-
 	!,
 	(	'$lgt_comp_ctx_mode'(Ctx, compile(_,_,_)),
@@ -11570,6 +11572,17 @@ create_logtalk_flag(Flag, Value, Options) :-
 	;	true
 	),
 	'$lgt_compile_body'(\+ Pred, TPred, DPred, Ctx).
+
+'$lgt_compile_body'(name(_, _), _, _, Ctx) :-
+	'$lgt_prolog_built_in_predicate'(name(_, _)),
+	\+ '$lgt_pp_defines_predicate_'(name(_, _), _, _, _, _, _),
+	'$lgt_comp_ctx_mode'(Ctx, compile(_,_,_)),
+	'$lgt_compiler_flag'(deprecated, warning),
+	'$lgt_source_file_context'(File, Lines),
+	'$lgt_pp_entity_'(Type, Entity, _),
+	'$lgt_increment_compiling_warnings_counter',
+	'$lgt_print_message'(warning(deprecated), deprecated_predicate(File, Lines, Type, Entity, name/2)),
+	fail.
 
 % warning on cuts on clauses for multifile predicates
 '$lgt_compile_body'(!, _, _, Ctx) :-